i2c-msm-v2: Ensure i2c bus is always active for transfers
There are quite a few i2c devices that are intended to remain active while
the system is suspended, which means that when one of them fires off an
interrupt, the subsequent i2c transfer inside the interrupt handler will
fail.

To fix this, make the system resume when an i2c transfer is requested
while the i2c bus is suspended. Additionally, add a 20 ms sleep after
calling the runtime PM resume function so that the hardware has enough
time to come up and we don't hit a chain of long i2c timeouts leading to
an eventual transfer failure.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>