oneplus5: defconfig: disable CONFIG_HARDEN_BRANCH_PREDICTOR
This setting was added along with a patch put in place to allow
hardening the branch predictor against aliasing attacks. To do this, the
CPU has to do additional work and its speculative capabilities might be
reduced, causing a dip in performance which could also affect day-to-day
use.
I'd rather keep it disabled, at least for the moment.

Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>