power: smb-lib: Fix charging issues due to incorrect intake voltage
The intake voltage with some charging sources goes out of range with the
current AICL solution. Instead of actually adjusting the current to get
back to a safe voltage, the charger driver currently sticks its head in
the sand and hopes everything will somehow sort itself out (which doesn't
happen).

Instead of doing nothing when the voltage is wrong, add a simple AICL
algorithm to adjust the current until the voltage is back to a safe
threshold. This fixes charging issues with some third party chargers.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>